Separated file uploader to method
